requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.8
cryptography==44.0.0
djangorestframework_simplejwt==5.4.0
drf-access-policy==1.5.0